refactor for lint
