+++ packages/ui-components/src/lib/wallet/asset.ts
@@ -1,9 +1,9 @@
 import config from '@unstoppabledomains/config';
 
-import type {TokenEntry} from '../../components/Wallet/Token';
 import {notifyEvent} from '../error';
 import {TokenType} from '../types';
 import type {AccountAsset} from '../types/fireBlocks';
+import type { TokenEntry} from '../types/wallet';
 import {SUPPORTED_SIGNING_SYMBOLS} from '../types/wallet';
 
 export const getAsset = (
